Crate scJaccard

scJaccard is a tool for quantifying the overlap between single cell peak calls and known purified cell types.

To use this tool, you need: - Peak calls in triplet MatrixMarket format (see here for details) - Known peaks in .bed format, sorted and index by tabix.

To run the tool, simply give the paths as arguments.

scJaccard --input   path/to/matrixMarket \
          --bed     path/to/file.bed \
          --atac    path/to/atac.bed

Input Formats

MatrixMarket.

Indexed bed.

Associated Tools

This tool has been created as a part of the avocato single cell Assay for Transposase-Accessible Chromatin using sequencing (scATAC-seq) pipeline.

Modules

utils

Structs

Cell

Opt

Record

Region

Functions

get_intersection

main

parse_bed_line

total_isec